moved upgrade test to a seperate suite
